Add jinja template functionality to template_render.py
